+++ base64.h
@@ -1,5 +1,6 @@
 /* ====================================================================
  * Copyright (c) 1995-1999 The Apache Group.  All rights reserved.
+ * 2017, Bernd Porr, mail@berndporr.me.uk
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -65,10 +66,10 @@ extern "C" {
 #endif
 
 int Base64encode_len(int len);
-int Base64encode(char * coded_dst, const char *plain_src,int len_plain_src);
+__int64 Base64encode(char * coded_dst, const char *plain_src,int len_plain_src);
 
-int Base64decode_len(const char * coded_src);
-int Base64decode(char * plain_dst, const char *coded_src);
+__int64 Base64decode_len(const char * coded_src);
+__int64 Base64decode(char * plain_dst, const char *coded_src);
 
 #ifdef __cplusplus
 }
